+++ client/web/src/auth/PostSignUpPage.tsx
@@ -20,7 +20,7 @@ import { UserExternalServicesOrRepositoriesUpdateProps } from '../util'
 import styles from './PostSignUpPage.module.scss'
 import { getReturnTo } from './SignInSignUpCommon'
 import signInSignUpCommonStyles from './SignInSignUpCommon.module.scss'
-import { Steps, Step, StepList, StepPanels, StepPanel, StepActions } from '../../../../wildcard/components'
+import { Steps, Step, StepList, StepPanels, StepPanel, StepActions } from '@sourcegraph/wildcard/src/components/Steps/components'
 import { useExternalServices } from './useExternalServices'
 import { CodeHostsConnection } from './welcome/CodeHostsConnection'
 import { Footer } from './welcome/Footer'
