+++ client/web/src/auth/welcome/StartSearching.tsx
@@ -9,7 +9,7 @@ import { eventLogger } from '../../tracking/eventLogger'
 import { UserExternalServicesOrRepositoriesUpdateProps } from '../../util'
 import { LogoAscii } from '../LogoAscii'
 import { getPostSignUpEvent, RepoSelectionMode } from '../PostSignUpPage'
-import { useSteps } from '../../../../../wildcard/components'
+import { useSteps } from '@sourcegraph/wildcard/src/components/Steps/components'
 import { Terminal, TerminalTitle, TerminalLine, TerminalDetails, TerminalProgress } from '../Terminal'
 import { useExternalServices } from '../useExternalServices'
 import { useRepoCloningStatus } from '../useRepoCloningStatus'
